refactor: extract sideNav components
